refactor: Result에서 GameResult로 클래스명 변경
